fixed integration
