+++ rex/rex_test.go
@@ -1,5 +1,1 @@
 package rex
-
-func GetQueueChan(ss *ReprocState) chan<- string {
-	return ss.queues
-}
